Reduce ping interval by 30 s, min. timeout: 45 s
The IPHB interval will be 30s, which means in the worst case it will
take 30s before we are waking up to send a PING to preempt
a server side timeout. Thus we trim _keepalive by 30 s when checking
if we need to send a PING.
This means PING messages will be sent within 30 s of the keepalive
target.
As a minimum keepalive 45 s seems like a nice value. 30 s would work
too, but then in the worst case we'd be sending PINGs every second
or so...